Resolve import conflicts in ffi.rs (add TiltFilterEffect, TubeCompressor
alongside DelayTiming) and Cargo.toml example ordering. Update delay
example Instrument trait to use f64 time signatures from upstream.

+++ examples/delay.rs
@@ -84,11 +84,11 @@ const PARAM_COUNT: usize = 5;
 struct SharedHiHat(Arc<Mutex<HiHat>>);
 
 impl Instrument for SharedHiHat {
-    fn trigger_with_velocity(&mut self, time: f32, velocity: f32) {
+    fn trigger_with_velocity(&mut self, time: f64, velocity: f32) {
         self.0.lock().unwrap().trigger_with_velocity(time, velocity);
     }
 
-    fn tick(&mut self, current_time: f32) -> f32 {
+    fn tick(&mut self, current_time: f64) -> f32 {
         self.0.lock().unwrap().tick(current_time)
     }
