+++ src/instruments/hihat.rs
@@ -509,14 +509,14 @@ impl HiHat {
     }
 
     /// Trigger with default velocity (1.0)
-    pub fn trigger(&mut self, time: f32) {
+    pub fn trigger(&mut self, time: f64) {
         self.trigger_with_velocity(time, 1.0);
     }
 
     /// Trigger with velocity sensitivity
     /// - High velocity: higher pitch, shorter decay, louder
     /// - Low velocity: lower pitch, longer decay, quieter
-    pub fn trigger_with_velocity(&mut self, time: f32, velocity: f32) {
+    pub fn trigger_with_velocity(&mut self, time: f64, velocity: f32) {
         self.is_active = true;
         self.current_velocity = velocity.clamp(0.0, 1.0);
 
@@ -599,7 +599,7 @@ impl HiHat {
         self.filter_envelope.trigger(time);
     }
 
-    pub fn release(&mut self, time: f32) {
+    pub fn release(&mut self, time: f64) {
         if self.is_active {
             self.noise_oscillator.release(time);
             self.brightness_oscillator.release(time);
@@ -608,7 +608,7 @@ impl HiHat {
         }
     }
 
-    pub fn tick(&mut self, current_time: f32) -> f32 {
+    pub fn tick(&mut self, current_time: f64) -> f32 {
         // Always tick smoothers (even when not active, to settle values)
         self.params.tick();
 
@@ -729,11 +729,11 @@ impl HiHat {
 
 // Implement the Instrument trait for engine compatibility
 impl crate::engine::Instrument for HiHat {
-    fn trigger_with_velocity(&mut self, time: f32, velocity: f32) {
+    fn trigger_with_velocity(&mut self, time: f64, velocity: f32) {
         HiHat::trigger_with_velocity(self, time, velocity);
     }
 
-    fn tick(&mut self, current_time: f32) -> f32 {
+    fn tick(&mut self, current_time: f64) -> f32 {
         self.tick(current_time)
     }
